+++ packages/prettier-plugin-sort-imports/src/types.ts
@@ -16,6 +16,8 @@ export interface ImportContent {
 
 /** 导入语句 */
 export interface ImportStatement {
+    /** 当前文件相对于当前路径的位置 */
+    filepath?: string
     /** 导入的模块路径，可以是相对路径或绝对路径，比如 react, react-dom 或者 ./utils/index，@/utils/index 等 */
     path: string
     /** 是否是导出语句，默认为 false */
@@ -40,6 +42,8 @@ export interface ImportStatement {
 
 /** 分组 */
 export interface Group {
+    /** 当前文件相对于当前路径的位置 */
+    filepath?: string
     /** 分组名称，默认为 default */
     name: string
     /** 是否是副作用分组，默认为 false */