+++ packages/prettier-plugin-remove-braces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @1adybug/prettier-plugin-remove-braces
 
+## 0.0.5
+
+### Patch Changes
+
+- 修复一些问题
+
 ## 0.0.4
 
 ### Patch Changes