+++ packages/prettier-plugin-sort-imports/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @1adybug/prettier-plugin-sort-imports
 
+## 0.0.19
+
+### Patch Changes
+
+- 修复一些问题
+
 ## 0.0.18
 
 ### Patch Changes